week5 functionMulti에서 프린트문의 서식지정자를 long로 통일했고 multif함수 원형 선언한 곳과 정의한 부분에서 반환형을 long로 통일함

+++ challenge/week5/functionMulti.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-int multif(int value);//multif 함수 원형 선언
+long multif(int value);//multif 함수 원형 선언
 
 int main() { //문자열 출력 및 함수 multif 호출
 	printf("1부터 2까지의 곱은 %ld\n", multif(2));
